refactor(app): clean lib.rs
